fix(geometry): split tritri.rs test module into tritri_tests.rs
The new interval-overlap regression test pushed tritri.rs to 421 lines,
past the module-size ratchet's 400-line budget with no allowlist row —
louistrue flagged this (verified on-branch: 421 -> 289 production +
tests file, sibling pattern already used by mesh_orient.rs/
mesh_orient_tests.rs and stream_meta.rs/stream_meta_tests.rs). Moving
the #[cfg(test)] mod tests block to a sibling file exempts it from the
ratchet without raising the budget or touching ALLOWLIST_DIGEST. No
behavior change.
